% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe.R
\name{as.data.frame.framespace}
\alias{as.data.frame.framespace}
\title{FrameSpace DataFrame to R DataFrame}
\usage{
\method{as.data.frame}{framespace}(dataframe)
}
\arguments{
\item{dataframe}{\strong{list} FrameSpace DataFrame}
}
\value{
R DataFrame
}
\description{
Converts a FrameSpace DataFrame into an R DataFrame
}

