% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe.R
\name{as.dataframe}
\alias{as.dataframe}
\title{FrameSpace DataFrame to R DataFrame}
\usage{
as.dataframe(dataframe)
}
\arguments{
\item{dataframe}{\strong{list} FrameSpace DataFrame}
}
\value{
R DataFrame
}
\description{
Converts a FrameSpace DataFrame into an R DataFrame
}

