% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axessearch.R
\name{axessearch}
\alias{axessearch}
\title{Sets up an axes search request}
\usage{
axessearch(url, names = NULL, pageSize = NULL, pageToken = NULL,
  raw = FALSE)
}
\arguments{
\item{url}{\strong{string} FrameSpace hostname or host:port to send request}

\item{names}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{\strong{integer}}

\item{pageToken}{\strong{string}}

\item{raw}{\strong{boolean} Flag to print raw or parsed content (default).}
}
\value{
FrameSpace SearchAxesResponse
}
\description{
Perform /axes/search request
}

