% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferreq.R
\name{bufferreq}
\alias{bufferreq}
\title{Buffers a dataframe slice request from the framespace server}
\usage{
bufferreq(host = "192.168.99.100", port = "5000",
  end = "/dataframe/slice", req = list(dataframeId =
  unbox("57912977105a6c0d293bbe8e")), buffer = 1000, looplimit = 1e+06)
}
\arguments{
\item{host}{= string This is the host you want the data to come from e.g. "192.168.99.100"}

\item{port}{= string containing the port you wish to access e.g. "5000"}

\item{end}{= string This is the prefex to the above url e.g. "example/search" (This is mostly for convenience sake and is not necisarry if the url section has everything in it)}

\item{req}{= list Input a list containing the search request structured as such e.g. list(dataframeId = unbox("57854014105a6c509d189029"), pageStart = unbox(37), pageEnd = unbox(500))}

\item{buffer}{= integer Imput the number of pages you would like to buffer for.}

\item{looplimit}{= integer The number of number of iterations you want the while loop to run for before you think it is stuck in an infinite loop. Defaults at 100,000.}
}
\value{
a list containig the full response
}
\description{
Beffers and sends a request in the form of a json file to the framespace server and retreaves the content of the response in packets.
}

