% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksetbuild.R
\name{buildKeySet}
\alias{buildKeySet}
\title{Construct a repeated set from provided query parameters}
\usage{
buildKeySet(url, keyspaceIds = NULL, keyspaceNames = NULL, keys = NULL)
}
\value{
repeated set for BuildDataFrameRequest
}
\description{
Construct a repeated set from query parameters
}

