% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbuild.R
\name{dfbuild}
\alias{dfbuild}
\title{Sets up a build dataframe request}
\usage{
dfbuild(url, majorSets, minorSets, units, pageSize = 0, raw = FALSE)
}
\arguments{
\item{majorSets}{\strong{string} keyspaceId for the major dimension}

\item{pageSize}{\strong{integer}}
}
\value{
FrameSpace DataFrame
}
\description{
Performs /dataframe request
}

