% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbuilder.R
\name{dfbuilder}
\alias{dfbuilder}
\title{Takes a}
\usage{
dfbuilder(host = "192.168.99.100", port = "5000",
  projectName = "tcga.BRCA", axisNames = "sample", newMajor = NULL,
  newMinor = NULL, id_num = 1, buffer = 1000, looplimit = 1e+05,
  df = TRUE)
}
\arguments{
\item{host}{\strong{string} The host you want the data to come from e.g. "192.168.99.100"}

\item{port}{\strong{string} The port you wish to access e.g. "5000"}

\item{projectName}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{axisNames}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{newMajor}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{newMinor}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{id_num}{\strong{int} Sometimes there will be multiple dataframes attached to one keyspace this identifies which one you want.}

\item{buffer}{\strong{integer} The number of pages you would like to buffer for.}

\item{looplimit}{\strong{integer} The number of times you want it to loop before it thinks it is in an infinite loop.}

\item{df}{\strong{bool} TRUE if you want the output in a dataframe format and FALSE if you want it in a matrix format.}
}
\value{
a dataframe/matrix with framespace data
}
\description{
Beffers and sends a request in the form of a json file to the framespace server and retreaves the content of the response in packets.
}

