% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsearch.R
\name{dfsearch}
\alias{dfsearch}
\title{Sets up a dataframe search request}
\usage{
dfsearch(keyspaceIds, dataframeIds = NULL, unitIds = NULL,
  pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{keyspaceIds}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{dataframeIds}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{unitIds}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{\strong{integer}}

\item{pageToken}{\strong{string}}
}
\value{
A properly formatted request
}
\description{
Formats a /dataframes/search request to send through the framereq function to the framespace server.
}

