% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genematrix.R
\name{genematrix}
\alias{genematrix}
\title{Converts a dataframe/slice response into a dataframe}
\usage{
genematrix(resp, df = TRUE)
}
\arguments{
\item{resp}{\strong{list} A responce from a dataframe/slice request}

\item{df}{\strong{bool} Leave TRUE if you want a dataframe or set to FALSE if you want a matrix}
}
\value{
A matrix structured with genes as rows and samples as collumns.
}
\description{
Converts a response from a dataframe/slice request into a dataframe or matrix stuctured with genes as collumns and samples as rows. This allows you to call on data like so: your_data[c("samplename1", "samplename2", "samplename3"), "genename"].
}

