% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genestat.R
\name{genestat}
\alias{genestat}
\title{Calculates means and variances for genes from a dataframe/slice response}
\usage{
genestat(resp, genes = NULL)
}
\arguments{
\item{resp}{\strong{list} A responce from a dataframe/slice request}

\item{genes}{\strong{vector} An atomic vector comtaining the names of genes that you want the mean and variance for.}
}
\value{
A list containing gene names associated with their respective means and variances.
}
\description{
Takes a responce from a dataframe/slice request and then calculates the mean and variance for specified genes. (or for all genes if the genes parameter is left NULL)
}

