% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataframe.R
\name{get.dataframe}
\alias{get.dataframe}
\title{Get DataFrame from KeySpace name and Unit name}
\usage{
get.dataframe(url, keyspaceName, unitName, pageStart = NULL, pageEnd = NULL,
  type = "data.frame")
}
\arguments{
\item{keyspaceName}{\strong{string} Name of the keyspace}

\item{unitName}{\strong{string} Name of the unit}

\item{pageStart}{\strong{integer}}

\item{pageEnd}{\strong{integer}}
}
\value{
R DataFrame
}
\description{
retrieves a dataframe from a given keyspace name and a unit name
}

