% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyidfind.R
\name{keyidfind}
\alias{keyidfind}
\title{Finds keyspace ids}
\usage{
keyidfind(resp)
}
\arguments{
\item{resp}{= a list from the output of a keyspace search using refreq}
}
\value{
An atomic vector containing keyspace ids.
}
\description{
This takes a response from a /keyspaces/search request then loads the ids into an atomic vector.
}

