% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpmethods.R
\name{post}
\alias{post}
\title{HTTP methods generalized for FrameSpace Requests}
\usage{
post(url, data)
}
\arguments{
\item{url}{\strong{string} FrameSpace hostname or host:port to send request}

\item{data}{\strong{list} list version of framespace proto request}
}
\value{
httr POST response
}
\description{
HTTP methods generalized for FrameSpace requests
}

