% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitssearch.R
\name{unitssearch}
\alias{unitssearch}
\title{Sets up a units search request}
\usage{
unitssearch(url, ids = NULL, names = NULL, pageSize = NULL,
  pageToken = NULL, raw = FALSE)
}
\arguments{
\item{ids}{\strong{vector} an atomic vector of strings e.g. c("a", "b", "c")}

\item{names}{\strong{vector} an atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{\strong{integer}}

\item{pageToken}{\strong{string}}

\item{raw}{\strong{boolean} Flag to print raw or parsed content (default).}
}
\value{
A properly formatted request
}
\description{
Formats a /units/search request to send through the framereq function to the framespace server.
}

