% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axessearch.R
\name{axessearch}
\alias{axessearch}
\title{Sets up an axes search request}
\usage{
axessearch(names = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{names}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{= integer}

\item{pageToken}{= string}
}
\value{
A properly formatted request
}
\description{
Formats an /axes/search request to send through the framefreq function to the framespace server.
}

