% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axessearch.R
\name{axessearch}
\alias{axessearch}
\title{Sets up an axes search request}
\usage{
axessearch(url, names = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{names}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{\strong{integer}}

\item{pageToken}{\strong{string}}
}
\value{
A properly formatted request
}
\description{
Formats an /axes/search request to send through the framefreq function to the framespace server.
}

