% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfslice.R
\name{dfslice}
\alias{dfslice}
\title{Sets up a dataframe slice request}
\usage{
dfslice(url, dataframeId, newMajorId = NULL, newMajorKeys = NULL,
  newMinorId = NULL, newMinorKeys = NULL, pageStart = NULL,
  pageEnd = NULL, raw = FALSE)
}
\arguments{
\item{dataframeId}{\strong{string}}

\item{newMajorId}{\strong{string} keyspaceId for the major dimension}

\item{newMajorKeys}{\strong{vector} atomic vector of keys associated with the major dimension}

\item{newMinorId}{\strong{string} keyspaceId for the minor dimension}

\item{pageStart}{\strong{integer}}

\item{pageEnd}{\strong{integer}}

\item{newMajorKeys}{\strong{vector} atomic vector of keys associated with the minor dimension}
}
\value{
FrameSpace DataFrame
}
\description{
Performs /dataframe/slice request
}

