% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfslice.R
\name{dfslice}
\alias{dfslice}
\title{Sets up a dataframe slice request}
\usage{
dfslice(dataframeId, newMajor = NULL, newMinor = NULL, pageStart = NULL,
  pageEnd = NULL)
}
\arguments{
\item{dataframeId}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{newMajor}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{newMinor}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{pageStart}{= integer}

\item{pageEnd}{= integer}
}
\value{
A properly formatted request
}
\description{
Formats a /dataframe/slice request to send through the framereq or bufferreq functions.
}

