% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genematrix.R
\name{genematrix}
\alias{genematrix}
\title{Converts a dataframe/slice response into a matrix}
\usage{
genematrix(resp)
}
\arguments{
\item{resp}{\strong{list} A responce from a dataframe/slice request}
}
\value{
A matrix structured with genes as rows and samples as collumns.
}
\description{
Converts a response from a dataframe/slice request into a matrix stuctured with genes as rows and samples as collumns. This allows you to call on data like so: your_data["genename", c("samplename1", "samplename2", "samplename3")].
}

