% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataframe.R
\name{get.dataframe}
\alias{get.dataframe}
\title{Get a dataframe for a given keyspace and unit.}
\usage{
get.dataframe(url, keyspaceName, unitName, pageStart = NULL, pageEnd = NULL,
  raw = FALSE)
}
\arguments{
\item{keyspaceName}{\strong{string} Name of the keyspace}

\item{unitName}{\strong{string} Name of the unit}

\item{pageStart}{\strong{integer}}

\item{pageEnd}{\strong{integer}}
}
\value{
A dataframe as an R datatable
}
\description{
retrieves a dataframe from a given keyspace name and a unit name
}

