% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysearch.R
\name{kssearch}
\alias{kssearch}
\title{Sets up a keyspaces search request}
\usage{
kssearch(url, axisNames, keyspaceIds = NULL, names = NULL, keys = NULL,
  pageSize = NULL, pageToken = NULL, raw = FALSE)
}
\arguments{
\item{axisNames}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{keyspaceIds}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{names}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{keys}{\strong{vector} An atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{\strong{integer}}

\item{pageToken}{\strong{string}}
}
\value{
A properly formatted request
}
\description{
Formats a /keyspaces/search request to send through the framereq function to the framespace server.
}

